Classes in Entity component folder moved to Doctrine folder
- classes are not related to Entities and these classes search in Doctrine
