+++ packages/framework/src/Resources/config/services.yml
@@ -591,9 +591,9 @@ services:
 
   Doctrine\Bundle\FixturesBundle\Loader\SymfonyFixturesLoader: '@doctrine.fixtures.loader'
 
-  Shopsys\FrameworkBundle\Component\Entity\EntityNotNullableColumnsFinder: ~
+  Shopsys\FrameworkBundle\Component\Doctrine\EntityNotNullableColumnsFinder: ~
 
-  Shopsys\FrameworkBundle\Component\Entity\EntityStringColumnsFinder: ~
+  Shopsys\FrameworkBundle\Component\Doctrine\EntityStringColumnsFinder: ~
 
   Shopsys\FrameworkBundle\Component\EntityExtension\EntityExtensionParentMetadataCleanerEventSubscriber:
     class: Shopsys\FrameworkBundle\Component\EntityExtension\EntityExtensionParentMetadataCleanerEventSubscriber
