+++ packages/framework/tests/Unit/Component/Doctrine/EntityNotNullableColumnsFinderTest.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Tests\FrameworkBundle\Unit\Component\Entity;
+namespace Tests\FrameworkBundle\Unit\Component\Doctrine;
 
 use Doctrine\Common\Persistence\Mapping\ClassMetadata;
 use Doctrine\ORM\Mapping\ClassMetadataInfo;
 use PHPUnit\Framework\TestCase;
-use Shopsys\FrameworkBundle\Component\Entity\EntityNotNullableColumnsFinder;
+use Shopsys\FrameworkBundle\Component\Doctrine\EntityNotNullableColumnsFinder;
 
 class EntityNotNullableColumnsFinderTest extends TestCase
 {
@@ -66,7 +66,7 @@ class EntityNotNullableColumnsFinderTest extends TestCase
     public function testGetAllNotNullableColumnNamesIndexedByTableNameException()
     {
         $classMetadataMock = $this->createMock(ClassMetadata::class);
-        $this->expectException(\Shopsys\FrameworkBundle\Component\Entity\Exception\UnexpectedTypeException::class);
+        $this->expectException(\Shopsys\FrameworkBundle\Component\Doctrine\Exception\UnexpectedTypeException::class);
 
         $entityNotNullableColumnsFinder = new EntityNotNullableColumnsFinder();
         $entityNotNullableColumnsFinder->getAllNotNullableColumnNamesIndexedByTableName([$classMetadataMock]);
